+++ lua/ui/lobby/lobby.lua
@@ -5956,9 +5956,16 @@ function InitHostUtils()
         -- Call this function once just before game starts
         SendArmySettingsToServer = function()
             local armyIdx = 1
-            for slotNum, playerInfo in gameInfo.PlayerOptions do
-                SendPlayerOption(playerInfo, 'Army', armyIdx)
-                armyIdx = armyIdx + 1
+            local MAXSLOT = 16
+            for slotNum = 1, MAXSLOT do
+                local playerInfo = gameInfo.PlayerOptions[slotNum]
+                if playerInfo ~= nil then
+                    LOG('>> SendArmySettingsToServer: Setting army '..armyIdx..' for slot '..slotNum)
+                    SendPlayerOption(playerInfo, 'Army', armyIdx)
+                    armyIdx = armyIdx + 1
+                else
+                    LOG('>> SendArmySettingsToServer: Slot '..slotNum..' empty')
+                end
             end
         end,